// src/services/tmdb/dtd/TMDB-dtd.d.ts
type TMDB_SizeDef =
    "w45" | "w92" | "w154" | "w185" | "w300" | "w342" | "w500" | "w780" | "w1280" |
    "h632" |
    "original";

type TMDB_ChangeKeyDef =
    "adult" | "air_date" | "also_known_as" | "alternative_titles" | "biography" | "birthday" | "budget" | "cast" | "certifications" |
    "character_names" | "created_by" | "crew" | "deathday" | "episode" | "episode_number" | "episode_run_time" | "freebase_id" | "freebase_mid" |
    "general" | "genres" | "guest_stars" | "homepage" | "images" | "imdb_id" | "languages" | "name" | "network" | "origin_country" | "original_name" |
    "original_title" | "overview" | "parts" | "place_of_birth" | "plot_keywords" | "production_code" | "production_companies" | "production_countries" |
    "releases" | "revenue" | "runtime" | "season" | "season_number" | "season_regular" | "spoken_languages" | "status" | "tagline" | "title" |
    "translations" | "tvdb_id" | "tvrage_id" | "type" | "video" | "videos";



export type CommonTMDBResultsResponse<T> = {
    page: number;
    results: T;
    total_pages: number;
    total_results: number;
}


export type GetTMDBConfigResponseDTD = {
    images: {
        base_url: string;
        secure_base_url: string;
        backdrop_sizes: TMDB_SizeDef[];
        logo_sizes: TMDB_SizeDef[];
        poster_sizes: TMDB_SizeDef[];
        profile_sizes: TMDB_SizeDef[];
        still_sizes: TMDB_SizeDef[];
    };
    change_keys: TMDB_ChangeKeyDef[];
}